refactore: CharacterState 분리
