+++ Assets/Scripts/Enums/CharacterState.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public enum CharacterState
+{
+    READY,
+    TURNCHECK,
+    WAITING,
+    ACTION
+}
