+++ Assets/Scripts/Enums/CharacterState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6112520e2fe7e4de7ad1efc525f5e530
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
