+++ Assets/Scripts/HeroController.cs
@@ -3,8 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
-
-public class MonsterState : MonoBehaviour
+public class HeroController : MonoBehaviour
 {
     private BattleManager battleManager;
     private Character character;
@@ -14,13 +13,6 @@ public class MonsterState : MonoBehaviour
     private bool isDead = false;
     private int spawnNumber; // 생성된 위치 넘버: 012 전열 345 후열
 
-    public enum CharacterState
-    {
-        TURNCHECK,
-        CHOOSEACTION,
-        WAITING,
-        ACTION
-    }
     public CharacterState currentState;
     public GameObject targetObject;
     public Animator animator;
@@ -31,18 +23,21 @@ public class MonsterState : MonoBehaviour
     {
         animator = GetComponent<Animator>();
         battleManager = GameObject.Find("Battle Manager").GetComponent<BattleManager>();
-        startPosition = transform.position;
-        currentState = CharacterState.TURNCHECK;
+        currentState = CharacterState.READY;
     }
 
     private void Update()
     {
         switch (currentState)
         {
+            case (CharacterState.READY):
+                StartCoroutine(TimeForReady());
+                currentState = CharacterState.TURNCHECK;
+                break;
             case (CharacterState.TURNCHECK):
-                if (battleManager.myTurn == true)
+                if (battleManager.myTurn == false)
                 {
-                    if (spawnNumber == battleManager.monsterNumber[0])
+                    if (spawnNumber == battleManager.heroNumber[0])
                     {
                         ChooseAction();
                         currentState = CharacterState.WAITING;
@@ -58,12 +53,29 @@ public class MonsterState : MonoBehaviour
         }
     }
 
+    private IEnumerator TimeForReady() 
+    {
+        yield return new WaitForSeconds(1f);
+        animator.SetTrigger("Forward");
+        while (MoveTowardBack()) { 
+            yield return null; 
+        }
+        animator.SetTrigger("Idle");
+        
+        yield return new WaitForSeconds(1f);
+        if (spawnNumber == battleManager.heroesInBattle.Count - 1) {
+            battleManager.GetReady();
+        }
+        yield break;
+    }
+
     void ChooseAction()
     {
         HandleTurn myAttack = new HandleTurn();
         myAttack.attackerGameObject = gameObject;
         battleManager.CollectActions(myAttack);
     }
+
 
     private IEnumerator TimeForAction()
     {
@@ -73,22 +85,15 @@ public class MonsterState : MonoBehaviour
         }
         actionStarted = true;
 
-        HeroState target = targetObject.GetComponent<HeroState>();
+        MonsterController target = targetObject.GetComponent<MonsterController>();
 
         animator.SetTrigger("Forward");
         while (MoveTowardEnemy()) { yield return null; }
 
-        // 공격 이펙트 실행과 데미지 계산
+        // TODO: attack 애니메이션 실행
         int damage = character.GetDamage();
-        if(damage == character.GetMaxDamage())
-        {
-            animator.SetTrigger("Critical");
-        }
-        else
-        {
-            animator.SetTrigger("Attack");
-
-        }
+        if (damage == character.GetMaxDamage()) animator.SetTrigger("Critical");
+        else animator.SetTrigger("Attack");
 
         GameObject effect;
         if (character.GetNature() == Nature.FIRE) effect = battleManager.fireEffect;
@@ -107,9 +112,9 @@ public class MonsterState : MonoBehaviour
         {
             target.animator.SetTrigger("Die");
             target.SetIsDead(true);
-            battleManager.heroesInBattle.Remove(targetObject);
-            battleManager.heroNumber.Remove(target.GetSpawnNumber());
-            battleManager.heroCps.Remove(target.GetCharacter().GetCP());
+            battleManager.monsterInBattle.Remove(targetObject);
+            battleManager.monsterNumber.Remove(target.GetSpawnNumber());
+            battleManager.monsterCps.Remove(target.GetCharacter().GetCP());
         }
         else
         {
@@ -123,12 +128,10 @@ public class MonsterState : MonoBehaviour
         animator.SetTrigger("Backward");
         while (MoveTowardBack()) { yield return null; }
 
-        // 제자리로 돌아와서 idle 애니메이션 실행
         animator.SetTrigger("Idle");
 
-        // BattleManager에 performList에서 하나를 제거
         battleManager.performList.RemoveAt(0);
-
+        
         actionStarted = false;
 
         // performList를 Wait로 reset
@@ -138,10 +141,9 @@ public class MonsterState : MonoBehaviour
     private bool MoveTowardEnemy()
     {
         Vector3 enemyPosition =
-            new Vector3(targetObject.transform.position.x, targetObject.transform.position.y, targetObject.transform.position.z - 2.5f);
+                    new Vector3(targetObject.transform.position.x, targetObject.transform.position.y, targetObject.transform.position.z + 2.5f);
         return enemyPosition != (transform.position = Vector3.MoveTowards(transform.position, enemyPosition, animateSpeed * Time.deltaTime));
     }
-
     private bool MoveTowardBack()
     {
         return startPosition != (transform.position = Vector3.MoveTowards(transform.position, startPosition, animateSpeed * Time.deltaTime));
@@ -158,7 +160,7 @@ public class MonsterState : MonoBehaviour
         healthSlider.maxValue = character.GetMaxHP();
         healthSlider.value = character.GetHP();
     }
-
+    
     public bool GetIsDead()
     {
         return isDead;
@@ -168,7 +170,7 @@ public class MonsterState : MonoBehaviour
     {
         isDead = value;
     }
-
+   
     public int GetSpawnNumber()
     {
         return spawnNumber;
@@ -177,5 +179,9 @@ public class MonsterState : MonoBehaviour
     public void SetSpawnNumber(int value)
     {
         spawnNumber = value;
+    }
+
+    public void SetStartPosition(Vector3 startPosition) {
+        this.startPosition = startPosition;
     }
 }
