+++ Assets/Scripts/MonsterSpawner.cs
@@ -17,9 +17,9 @@ public class MonsterSpawner : MonoBehaviour
 
             // 유닛 생성
             GameObject spawnUnit = Instantiate(prefab, transform.position, Quaternion.identity);
-            spawnUnit.GetComponent<MonsterState>().SetCharacter(spawnCharacter);
-            spawnUnit.GetComponent<MonsterState>().SetSpawnNumber(locationNumber);
-            spawnUnit.GetComponent<MonsterState>().gameObject.SetActive(true);
+            spawnUnit.GetComponent<MonsterController>().SetCharacter(spawnCharacter);
+            spawnUnit.GetComponent<MonsterController>().SetSpawnNumber(locationNumber);
+            spawnUnit.GetComponent<MonsterController>().gameObject.SetActive(true);
 
             battleManager = GameObject.Find("Battle Manager").GetComponent<BattleManager>();
             battleManager.monsterInBattle.Add(spawnUnit);
